% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare.DCIM}
\alias{compare.DCIM}
\title{Compare folders containing qfield attachments}
\usage{
compare.DCIM(path_qfield = NULL, path_local = NULL)
}
\arguments{
\item{path_qfield}{path to QFieldCloud-Project}

\item{path_local}{path to a local copy of attachments of the same QFieldCloud-Project}
}
\description{
To save space in a QFieldCloud-Project it can help to reduce image sizes by replacing original file versions by a thumbnail. This function allows to quickly identify new files and check for deleted images in the cloud.
}
