% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_raster}
\alias{download_raster}
\title{Download files}
\usage{
download_raster(url, filename = basename(url), dest)
}
\arguments{
\item{url}{url}

\item{filename}{name}

\item{dest}{output folder}
}
\description{
Small function to download raster files (e.g. tif, png) in a batch mode
}
