% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shrink_gpkg}
\alias{shrink_gpkg}
\title{Reduce GeoPackage size to save disk space}
\usage{
shrink_gpkg(dsn = NULL, backup = NULL)
}
\arguments{
\item{dsn}{Path to GeoPackage}

\item{backup}{optional create backup before running VACUUM}
}
\description{
Running VACUUM on a GeoPackage
}
