% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shrink_gpkg_batch}
\alias{shrink_gpkg_batch}
\title{Batch apply \link{shrink_gpkg} to all geopackage files within a folder}
\usage{
shrink_gpkg_batch(path = NULL, recursive = FALSE)
}
\arguments{
\item{path}{path}

\item{recursive}{boolean}
}
\description{
See \link{shrink_gpkg} for details
}
