% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sync}
\alias{sync}
\title{clean database by removing duplicated entries based on [uuid](https://en.wikipedia.org/wiki/Universally_unique_identifier) and MTIME}
\usage{
sync(DF = NULL, UUID = "uuid", MTIME = "MTIME")
}
\arguments{
\item{DF}{data frame after merging file versions}

\item{UUID}{string denoting field representing an universally unique identifier [uuid](https://en.wikipedia.org/wiki/Universally_unique_identifier)}

\item{MTIME}{string denoting field representing a timestamp of last modification}
}
\description{
clean database by removing duplicated entries based on [uuid](https://en.wikipedia.org/wiki/Universally_unique_identifier) and MTIME
}
