% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{VACUUM.batch}
\alias{VACUUM.batch}
\title{Batch apply \link{VACUUM} to all geopackage files within a folder}
\usage{
VACUUM.batch(path = NULL, recursive = FALSE)
}
\arguments{
\item{path}{path}

\item{recursive}{boolean}
}
\description{
See \link{VACUUM} for details
}
