% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{VACUUM}
\alias{VACUUM}
\title{Reduce GeoPackage size to save disk space}
\usage{
VACUUM(dsn = NULL, backup = NULL)
}
\arguments{
\item{dsn}{Path to GeoPackage}

\item{backup}{optional create backup before running VACUUM}
}
\description{
Running VACUUM on a GeoPackage
}
